+++ src/types/dtos.ts
@@ -77,13 +77,11 @@ export interface User {
   userId: number,
   objectId: string,
   isActive: boolean,
-  organisation: {
-    id: number,
-    abbreviation: string,
-  }
+  orgId: number,
+  orgAbbrev: string,
   orgName: string,
   isAustrakkaAdmin: boolean,
-  userRoleGroup: UserRoleGroup[],
+  groupRoles: GroupRole[],
   displayName: string,
   createdBy: Date,
   contactEmail: string,
@@ -94,7 +92,8 @@ export interface User {
 export interface UserDetails {
   displayName: string,
   orgName: string,
-  userRoleGroup: UserRoleGroup[],
+  orgAbbrev: string,
+  groupRoles: GroupRole[],
   created: Date,
   contactEmail: string,
 }
@@ -214,8 +213,7 @@ export interface ThresholdAlertDTO {
   recentCount: number;
 }
 
-export interface UserRoleGroup {
-  user: any,
+export interface GroupRole {
   role: {
     id: number,
     name: string,